refactor: remove `useDocuments` and `useSearch`
