+++ packages/core/src/document/documentStore.ts
@@ -26,14 +26,13 @@ import {
 } from 'rxjs'
 
 import {getClientState} from '../client/clientStore'
-import {API_VERSION} from '../documentList/documentListConstants'
 import {type SanityInstance} from '../instance/types'
 import {type ActionContext, createAction, createInternalAction} from '../resources/createAction'
 import {createResource} from '../resources/createResource'
 import {createStateSourceAction, type StateSource} from '../resources/createStateSourceAction'
 import {getDraftId} from '../utils/ids'
 import {type DocumentAction} from './actions'
-import {INITIAL_OUTGOING_THROTTLE_TIME} from './documentConstants'
+import {API_VERSION, INITIAL_OUTGOING_THROTTLE_TIME} from './documentConstants'
 import {type DocumentEvent, getDocumentEvents} from './events'
 import {listen, OutOfSyncError} from './listen'
 import {type DocumentHandle, type JsonMatch, jsonMatch, type JsonMatchPath} from './patchOperations'
