--- packages/core/src/documentList/documentListConstants.ts
@@ -1,2 +1,0 @@
-export const API_VERSION = 'vX'
-export const PAGE_SIZE = 25
