+++ packages/react/src/_exports/hooks.ts
@@ -25,9 +25,6 @@ export {useDocumentEvent} from '../hooks/document/useDocumentEvent'
 export {useDocumentSyncStatus} from '../hooks/document/useDocumentSyncStatus'
 export {useEditDocument} from '../hooks/document/useEditDocument'
 export {usePermissions} from '../hooks/document/usePermissions'
-export {type DocumentHandleCollection} from '../hooks/documentCollection/types'
-export {useDocuments} from '../hooks/documentCollection/useDocuments'
-export {useSearch} from '../hooks/documentCollection/useSearch'
 export {
   type InfiniteList,
   type InfiniteListOptions,
